Add files via upload
I was able to fetch the data on given API using axios command however the same data is responding to be written in HTML as it is a json format data. I used $each method to traverse through the data but it doesn't seem  to work giving me no output. I don't know where is this code missed!